Add missing Bazel-Gazelle workspace declaration in TECL's WORKSPACE
This is used to allow Gazelle managing go dependencies. Since go.mod and
go.sum are copied (managed) by CopyBara, this aligns the ability to sync
TECL's workspace too on changes. For more information see the error message
Gazelle produces when run on this repository:
https://github.com/bazelbuild/bazel-gazelle/blob/071c8d00482fc8e4dc65c451c5d41411631280c1/merger/fix.go#L172-L178

This was omitted in the bump of Gazelle to 0.18.1 in Change-Id
I747f2addfe9f20c726fda271fded4bac1d2b5ca3 (commit
3b28e810f0c6c89a6c2577f3f803496686ed99fd in TECL).

Change-Id: I8d40ce01a08088eb354643600d142ff06e01e252
GitOrigin-RevId: e90a8363ceade583d0517bae76b2739c6257be8e