Update Gazelle to latest master and bazelbuild/buildtools (Buildifier)
This has to be a single commit, because custom Gazelle rules in this
repository create a cross-dependency due to a change in
bazelbuild/buildtools, which required a change in Gazelle (unreleased as of
yet).

Gazelle does not want to have empty BULID.bazel files for what it does not
concern itself as a package. So, as a work-around, they now have a comment,
as dummy content. See also:
https://github.com/bazelbuild/bazel-gazelle/issues/853

Change-Id: I16936d6a8ef665706f207833c491a1dc4eba7f1a
GitOrigin-RevId: 3106fe661d57023d7cb3b92b7c71258e197edcc4